Add protobuf_catkin, find_package ompl, tmp_byte_offset to uint64

+++ voxblox_skeleton/src/io/skeleton_io.cpp
@@ -66,7 +66,7 @@ bool loadSparseSkeletonGraphFromFile(const std::string& filename,
 
   // Byte offset result, used to keep track where we are in the file if
   // necessary.
-  uint32_t tmp_byte_offset = 0;
+  uint64_t tmp_byte_offset = 0; // TODO: Fork orig repo, push change and submit pull request
 
   SkeletonGraphProto proto;
   if (!utils::readProtoMsgFromStream(&proto_file, &proto, &tmp_byte_offset)) {
